Split out the Atrac context class into AtracCtx.cpp/h
