+++ Core/Core.vcxproj.filters
@@ -1309,6 +1309,9 @@
     <ClCompile Include="HW\Atrac3Standalone.cpp">
       <Filter>HW</Filter>
     </ClCompile>
+    <ClCompile Include="HLE\AtracCtx.cpp">
+      <Filter>HLE\Libraries</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="ELF\ElfReader.h">
@@ -2093,6 +2096,9 @@
     </ClInclude>
     <ClInclude Include="HW\Atrac3Standalone.h">
       <Filter>HW</Filter>
+    </ClInclude>
+    <ClInclude Include="HLE\AtracCtx.h">
+      <Filter>HLE\Libraries</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
